Allow metrics searches in addition to hosts
This is basically the same code, just with a different facet from the
API's point of view, so hosts/metrics/searching all got split up to
allow working with both.

This raises some concerns about the Utilities module being included in
the handler, so I want to rip that out at some point to clean up
things.